requirements.txt
Django<4
djangorestframework>=3.11.0,<3.12.0
django-cors-headers==3.2.1
google-generativeai